pyproject.toml
[tool.poetry]
name = "mafia"
version = "0.1.0"
description = "mafia game for holbies"
authors = ["VoiceOfDarkness <abil.samedov2016@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
flask = "^3.0.3"
jinja2 = "^3.1.4"
python-dotenv = "^1.0.1"
icecream = "^2.1.3"
gunicorn = "^22.0.0"
flask-pymongo = "^2.3.0"
isort = "^5.13.2"
flask-socketio = "^5.3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
